Update README.md
I added one line in the "Option 2: In a shared folder" installation guide to counter the following issue:

The "Integrity Excel Test Session" window does not open and no error (only a successful connection) is logged when using the "Option 2: In a shared folder" installation.

From the MANIFEST file in the IntegrityExcelTestSession.jar it becomes clear that "mksclient.jar" is also a needed library, but it is not included in the dist/lib folder. For this reason a workaround is to copy the "mksclient.jar" file from the Integrity installation to the dist/lib folder. For the sake of completeness I also replaced the "mksapi.jar" in the dist/lib folder with the one from the Integrity installation. This was tested under Integrity 11.1 and I do not know if it applies to other versions as well.